Lilys-guessing-game
Project title: Lilys Guessing Game. 

The game generates a random int each time it is started at the end of the program that random int is reset. The game asks for a name and if the user types close at any point of the input situations the game exits.  The games keeps track  of guesses the user makes. The game imports the instances random and sys in order to create the random number and in order to exit the system. The game provides feedback to the user  by using a conditional which is "if, elif, elif" in order to tell the user if their guess  is too high, too low. After the guess is correct it adds the username and score to a file named "score.txt". The score is later printed into the console. If the input is incorrect the user is reminded of what is the correct input to continue. The user is asked if they want to continue playing otherwise they exit the system. All the code is written in main and doc-strings are added (but I might have missed some because I am noob). 

Motivation:
The program was created with the intention to make me less of a noob.

Builds status:
There (hopefully) are no bug /error which needs addressing.

Code Style:
Standard.

Tech/Framework used:
Python. 

How to Use?

https://www.wikihow.com/Use-Windows-Command-Prompt-to-Run-a-Python-File

Installation:
Anaconda for test environment. CMD.exe.

Credits:
Google.com.